[feature] Add Bazel run line markers in Go files
